fix int bonus

+++ src/BonusMilesService.java
@@ -1,7 +1,7 @@
 public class BonusMilesService {
     public int calculate(int cost) {
         int coins = 20;
-        long bonus = cost / coins;
+        int bonus = cost / coins;
 
         return (int) bonus;
     }
